app/requirements.txt
alembic==1.8.1 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.9" and python_version < "4.0"
asyncpg==0.26.0 ; python_version >= "3.9" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.9.24 ; python_version >= "3.9" and python_version < "4"
cffi==1.15.1 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==2.1.1 ; python_version >= "3.9" and python_version < "4"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
cryptography==38.0.3 ; python_version >= "3.9" and python_version < "4.0"
databases==0.6.1 ; python_version >= "3.9" and python_version < "4.0"
dnspython==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.9" and python_version < "4.0"
email-validator==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
fastapi==0.85.2 ; python_version >= "3.9" and python_version < "4.0"
greenlet==2.0.1 ; python_version >= "3.9" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
mako==1.2.3 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.1 ; python_version >= "3.9" and python_version < "4.0"
passlib[bcrypt]==1.7.4 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.2 ; python_version >= "3.9" and python_version < "4.0"
python-decouple==3.6 ; python_version >= "3.9" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.5 ; python_version >= "3.9" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.9" and python_version < "4"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
sqlalchemy==1.4.43 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.20.4 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.12 ; python_version >= "3.9" and python_version < "4"
uvicorn==0.18.3 ; python_version >= "3.9" and python_version < "4.0"
